+++ installation-scripts/install_docker.sh
@@ -1,44 +1,50 @@
 #!/bin/bash
 
+set +e
+
 #Run the following script in root mode
 
 # Look for processes holding locks.
 # If no process is holding the lock then remove the lock.
-declare -a locks=("/var/lib/dpkg/lock*" "/var/lib/apt/lists/lock" "/var/cache/apt/archives/lock")
+
+declare -a locks=("/var/lib/dpkg/lock" "/var/lib/dpkg/lock-frontend" "/var/lib/apt/lists/lock" "/var/cache/apt/archives/lock")
 FLAG=0
-for val in ${locks[@]}; do
-  PID=$(lsof -t $val 2>&1)
-  LSOF_EXIT_CODE=$?
-  WAIT_TIME=2
-
-  while [ $LSOF_EXIT_CODE -eq 0 ] && [[ $PID =~ ^[0-9]+$ ]] && [ $WAIT_TIME -gt 0 ]
-  do
-    echo
-    echo Lock: $val held by process $PID
-    echo Waiting $[WAIT_TIME*10] seconds for process $PID to exit..
-    sleep 10
-    WAIT_TIME=$[WAIT_TIME-1]
-    PID=$(lsof -t $val 2>&1)
-    LSOF_EXIT_CODE=$?
-  done
-
-  if [ $WAIT_TIME -le 0 ] || [[ $PID =~ ^[0-9]+$ ]]
+for val in "${locks[@]}"
+do
+  if [ -f "$val" ]
   then
-    echo
-    echo Process $PID is still running with lock $val held.
-    echo Terminating Docker CE installation. 
-    echo Exiting with code=1
-    exit 1
-  fi
+	PID=`lsof -t "$val"`
+  	LSOF_EXIT_CODE=$?
+  	WAIT_TIME=2
+  	while [ $LSOF_EXIT_CODE -eq 0 ] && [[ $PID =~ ^[0-9]+$ ]] && [ $WAIT_TIME -gt 0 ]
+  	do
+		echo
+   		echo Lock: $val held by process $PID
+    		echo Note: Installation will be paused until the locks are realeased
+    		echo Waiting $[WAIT_TIME*10] seconds for process $PID to exit..
+    		sleep 10
+    		WAIT_TIME=$[WAIT_TIME-1]
+    		PID=$(lsof -t $val 2>&1)
+    		LSOF_EXIT_CODE=$?
+  	done
 
-  ls -l $val > /dev/null 2>&1
-  if [ $? -eq 0  ]
-  then
-    echo No process is holding $val. Releasing it..
-    rm -rf $val
-    FLAG=1
-  fi
+  	if [ $WAIT_TIME -le 0 ] || [[ $PID =~ ^[0-9]+$ ]]
+  	then
+    		echo
+    		echo Process $PID is still running with lock $val held.
+    		echo Terminating Docker CE installation. 
+    		echo Exiting with code=1
+    		exit 1
+  	fi
 
+  	ls -l $val > /dev/null 2>&1
+  	if [ $? -eq 0  ]
+  	then
+    		echo No process is holding $val. Releasing it..
+    		rm -rf $val
+    		FLAG=1
+  	fi
+  fi
 done
 
 # If any locks removed reconfigure the packages
@@ -50,26 +56,45 @@ then
 fi
 
 # Remove older versions of docker
-apt-get remove docker docker-engine docker.io containerd runc
+echo 
+echo Removing old Docker installations if found any..
+echo
+apt-get -y remove docker docker-engine docker.io containerd runc
 
 
 # Catch errors from here
 set -e
 
-# Set Proxy Variables
-# To set environment variables through script use 'source' to build
-IP=$(awk -F'proxy_ip:' '{print $2}' .gitignore)
-PORT=$(awk -F'proxy_port:' '{getline;print $2}' .gitignore)
-PROXY="$IP:$PORT"
-export https_proxy=$PROXY
-export http_proxy=$PROXY
+if [ $PROXY_FLAG -eq 1 ]
+then
+	# Set Proxy Variables
+	# To set environment variables through script use 'source' to build
+	IP=$(awk -F'proxy_ip:' '{print $2}' $config_file)
+	PORT=$(awk -F'proxy_port:' '{getline;print $2}' $config_file)
+	PROXY="$IP:$PORT"
 
+	echo
+	echo "Parsing proxy address.."
+	echo "Parsed proxy address: $PROXY"
+	echo
+	echo Exporting proxies..
+
+	export https_proxy=$PROXY
+	export http_proxy=$PROXY
+
+else
+	echo 
+	echo No Proxies Set
+fi
 
 # Install Docker CE from repository
-apt update
-apt upgrade
+echo 
+echo "Updating packages.."
+echo
+apt-get -y update
+apt-get -y upgrade
 
-apt-get install \
+apt-get -y --force-yes install \
     apt-transport-https \
     ca-certificates \
     curl \
@@ -85,10 +110,13 @@ add-apt-repository \
    $(lsb_release -cs) \
    stable"
 
-apt update
-apt upgrade
+apt-get -y update
+apt-get -y upgrade
 
-apt-get install docker-ce docker-ce-cli containerd.io
+echo
+echo Installing Docker, Docker-ce/ce-cli, containerd.io
+echo
+apt-get -y --force-yes install docker-ce docker-ce-cli containerd.io
 
 
 # Verify docker installation
@@ -102,4 +130,5 @@ else
     echo
     echo "Docker Installation Failed"
     echo
+    exit 1
 fi
